# specs/001-document-qa-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant Document Q&A API
  version: 1.0.0
  description: Asynchronous REST API for user authentication, document management, and chat history.
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Email already registered or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      summary: Authenticate user and obtain JWT token
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      summary: Get current authenticated user profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized

  /documents:
    get:
      summary: List all documents owned by the authenticated user
      operationId: listUserDocuments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
        '401':
          description: Unauthorized
    post:
      summary: Upload a new document (PDF or TXT) for background processing
      operationId: uploadDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF or plain text file (max 25MB)
              required:
                - file
      responses:
        '202':
          description: File accepted and background processing initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid file type or empty file
        '413':
          description: File exceeds 25MB limit

  /documents/{document_id}:
    get:
      summary: Get status and metadata of a specific document
      operationId: getDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found or not owned by user
    delete:
      summary: Delete a document and purge all associated vector chunks
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document and chunks permanently deleted
        '404':
          description: Document not found or not owned by user

  /chat/sessions:
    get:
      summary: List all chat sessions for the authenticated user
      operationId: listChatSessions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSessionResponse'
    post:
      summary: Create a new chat session
      operationId: createChatSession
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  default: New Conversation
      responses:
        '201':
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionResponse'

  /chat/sessions/{session_id}:
    get:
      summary: Get chat session details and message history
      operationId: getChatSession
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chat session and historical messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetailResponse'
        '404':
          description: Session not found or not owned by user
    patch:
      summary: Rename chat session title
      operationId: updateChatSession
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
              required:
                - title
      responses:
        '200':
          description: Updated chat session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionResponse'
    delete:
      summary: Delete chat session and its messages
      operationId: deleteChatSession
      security:
        - BearerAuth: []
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session deleted
        '404':
          description: Session not found

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required:
        - access_token
        - token_type
        - user
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        user:
          $ref: '#/components/schemas/UserResponse'

    UserResponse:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    DocumentResponse:
      type: object
      required:
        - id
        - filename
        - file_size
        - mime_type
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        status:
          type: string
          enum: [uploading, processing, ready, failed]
        total_pages:
          type: integer
          nullable: true
        error_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    ChatSessionResponse:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    MessageResponse:
      type: object
      required:
        - id
        - role
        - content
        - created_at
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        citations:
          type: array
          items:
            type: object
            properties:
              document_id:
                type: string
                format: uuid
              document_name:
                type: string
              page_number:
                type: integer
              snippet:
                type: string
        created_at:
          type: string
          format: date-time

    ChatSessionDetailResponse:
      allOf:
        - $ref: '#/components/schemas/ChatSessionResponse'
        - type: object
          properties:
            messages:
              type: array
              items:
                $ref: '#/components/schemas/MessageResponse'

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string